create separate dao and service for guest-api
