+++ src/main/java/com/rentmycar/controller/GuestController.java
@@ -1,7 +1,6 @@
 package com.rentmycar.controller;
 
 import javax.validation.constraints.Max;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,8 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.rentmycar.service.UserService;
-
+import com.rentmycar.service.GuestService;
 import io.swagger.v3.oas.annotations.Operation;
 
 @RestController
@@ -18,12 +16,12 @@ import io.swagger.v3.oas.annotations.Operation;
 public class GuestController {
 
 	@Autowired
-	private UserService userService;
+	private GuestService guestService;
 	
 	@GetMapping("/profile/{guestId}")
 	@Operation(description = "getProfileDetails By Id")
 	public ResponseEntity<?> getGuestProfileDetails(@PathVariable @Max(100) Long guestId){
 		System.out.println("In Guest Profile" + guestId);		
-		return ResponseEntity.ok(userService.getGuestProfileDetails(guestId));
+		return ResponseEntity.ok(guestService.getGuestProfileDetails(guestId));
 	}
 }
