+++ src/main/java/com/rentmycar/service/GuestService.java
@@ -1,0 +1,10 @@
+package com.rentmycar.service;
+
+import java.util.Optional;
+
+import com.rentmycar.dto.GuestDetailsResponseDto;
+
+public interface GuestService {
+
+	Optional<GuestDetailsResponseDto> getGuestProfileDetails(Long guestId);
+}
