+++ src/main/java/com/rentmycar/service/UserService.java
@@ -1,8 +1,6 @@
 package com.rentmycar.service;
 
 import java.util.Optional;
-
-import com.rentmycar.dto.GuestDetailsResponseDto;
 import com.rentmycar.dto.SignInRequestDto;
 import com.rentmycar.dto.SignInResponseDto;
 
@@ -10,5 +8,4 @@ public interface UserService {
 	
 	Optional<SignInResponseDto> authenticateUser(SignInRequestDto request);
 	
-	Optional<GuestDetailsResponseDto> getGuestProfileDetails(Long guestId);
 }
